Bugfix
A bug is detected in the path finder part. When,
i. there are multiple successive switches: sw0, sw1, sw2
ii. connections:  es0-swo-sw1-es1, es0-sw0-sw2-es2,
iii. there is a vl like this: es0->es1,es2.
Lets say sw1.port2: es1-sw1 and sw1port9:sw1-sw0.
In that case it is expected to see in the sw1 configuration table vl1 : {2}. But since the frames of vl1 are also directed to es2, the resulting configuration table was like this: vl1: {2,9} Which is faulty. A behavior like this would cause packets get multiplies more and more and lower the bandwidth. So an additional control is added for this bug